backend/requirements.txt
Flask
flask-cors
requests
Pillow
gunicorn